Test out js functions
