+++ app/assets/javascripts/application.js
@@ -16,6 +16,7 @@
 //= require_tree .
 //= require 'colorFunctions'
 //= require 'musicFunctions'
+//= require 'keypress.js'
 //= require jquery-ui/effect-shake
 var div = ["7","8","9","4","5","6","1","2","3"];
 var loadColors = ["#eafaf1","#d5f5e3","#abebc6","#82e0aa","#58d68d","#2ecc71","#28b463","#239b56","#1d8348"];
@@ -28,27 +29,15 @@ $(document).ready(function() {
     // setTimeout($("#key-"+div[i]).css('background-color', 'black'), 1000);
   });
 
-  // Return square to black when key is depressed
-  $('body').keyup(function(event) {
-    var element_id = '#key-' + (event.keyCode - 48).toString();
-    // add delay
-    setTimeout(function(){
-      $(element_id).css('background-color', 'black');
-    }, 3000);
-  })
+  // $("#key-7, #key-8, #key-9, #key-6, #key-3, #key-2, #key-1, #key-4, #key-5").hide().each(function(i) {
+  //   $(this).css('background-color', 'black').delay(i*400)
+  //   // setTimeout($("#key-"+div[i]).css('background-color', 'black'), 1000);
+  // });
 
-  // Change box color when key is pressed
-  $('body').keydown(function(event) {
-    var key_code = event.keyCode;
-    var color = randomColor();
-    playKeypress(key_code, color);
+  // $("7, 8, #key-9, #key-6, #key-3, #key-2, #key-1, #key-4, #key-5").hide().each(function(i) {
+  //   $("#key-"+this).delay(i*400).fadeIn(200);
+  // });
 
-    if (recording) {
-      this_press_timestamp = new Date().getTime();
-      recordBeat(key_code, last_press_timestamp, this_press_timestamp, color);
-      last_press_timestamp = this_press_timestamp;
-    }
-  })
 
   $('#play-track').click(function(event) {
     event.preventDefault();
@@ -81,14 +70,9 @@ $(document).ready(function() {
 
     // send track back to DB (an array of layer objects).tracks is the recording
 
-    // ask for name here?
-    var name = prompt("Your track is lonely! Give it a name.")
-
-    console.log(name)
-
     $.ajax({
       url: '/tracks',
-      data: {'track': tracks, 'name': name},
+      data: {'track': tracks},
       method: 'post'
     })
   })
@@ -103,11 +87,9 @@ $(document).ready(function() {
     var id_of_track_to_play = play_button_clicked.attr('id')
     console.log(play_button_clicked.attr('id'));
 
-
-
-
     // find contents of track to play, the array
     // var contents_of_track_to_play =
+
 
     // LASTLY:
     // playTracks( track.contents );
